tournament cycle actually completed. Probably still has bugs in results and such, but no errors are being shown anymore

+++ testBackendFINAL/services/TournamentService.ts
@@ -129,6 +129,7 @@ class TournamentService {
         console.log(`Match with ID ${matchID} just finished with the following result:`);
         console.log(result);
         const matchUsers = matchRepository.getMatchUsersByID(matchID);
+        if (!matchUsers) {throw Error("This match does not exist!")}
         socketService.broadcastToUsers(matchUsers, "endOfMatch", { result: result });
         matchService.destroyMatchByID(matchID);
